exercise 4.21*: Blog List Expansion, step 9

+++ part4/blogList-backend/controllers/blogs.js
@@ -30,7 +30,6 @@ blogsRouter.post('/', async (request, response) => {
   }
 
   const user = await User.findById(decodedToken.id)
-  console.log('user', user);
 
   if (!body.title || !body.url) {
     return response.status(400).json({
@@ -80,6 +79,24 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    
+  const decodedToken = jwt.verify(request.token, config.SECRET_TOKEN)
+  console.log('decodedToken',decodedToken);
+
+  if(!decodedToken.id){
+    return response.status(401).json({error:'invalid token'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if(!blog){
+    return response.status(404).json({error:'blog not found!'})
+  }
+
+  if(blog.user.toString() !== decodedToken.id.toString()){
+    return response.status(401).json({error: 'only the creator can delete this blog'})
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
